refractor code

+++ scripts/synthentic_validating_data.py
@@ -4,9 +4,10 @@ import json
 from ig_rewarding import Validator
 import pandas as pd
 import json
-from typing import List
-from tqdm import tqdm
+from typing import List, Dict
+import tqdm
 import numpy as np
+import easydict
 from transformers import Pipeline
 from diffusers import StableDiffusionPipeline, StableDiffusionXLPipeline
 import yaml
@@ -15,16 +16,19 @@ import seaborn as sns
 import torch
 from ig_rewarding.utils import instantiate_from_config
 from argparse import ArgumentParser
+
 synthentic_config_file = "scripts/synthentic_config.yaml"
 config_file = "ig_rewarding/config/baseline.yaml"
 config = yaml.load(open(config_file, "r"), Loader=yaml.FullLoader)
 synthentic_config = yaml.load(open(synthentic_config_file, "r"), Loader=yaml.FullLoader)
+synthentic_config = easydict.EasyDict(synthentic_config)
+
 
 def validate_pipe(
     validator: Validator,
     pipe: Pipeline,
     pipeline_name: str,
-    topics: List[str],
+    topics: Dict[str, str],
     n_try: int,
     n_image_per_prompt: int,
     save_dir: str = "images",
@@ -34,28 +38,30 @@ def validate_pipe(
     raw_pipeline_name = pipeline_name
     pipeline_name = pipeline_name.replace("/", "_")
     request_result = []
+    topic_names = list(topics.keys())
+    topic_prefixes = [topics[topic_name]["prefix"] for topic_name in topic_names]
+    all_topic_prompts = validator.generate_prompt(topic_prefixes, n_image_per_prompt)
+
     for i in range(n_try):
         request_id = i
-        prompts = validator.generate_prompt(topics)
-
-        for j in range(n_image_per_prompt):
-            images = []
-            for prompt in prompts:
-                outputs = pipe(
-                    [prompt],
-                    num_images_per_prompt=1,
-                    negative_prompt=[negative_prompt],
-                )
-                images.append(outputs.images[0])
+        prompts = [topic_prompts[i] for topic_prompts in all_topic_prompts]
+        print(prompts)
+        for j, prompt in enumerate(prompts):
+            outputs = pipe(
+                [prompt],
+                num_images_per_prompt=n_image_per_prompt,
+                negative_prompt=[negative_prompt],
+            )
+            images = outputs.images
             file_names = [
-                f"{pipeline_name}_{i}_{j*len(prompts)+t}.webp"
-                for t in range(len(prompts))
+                f"{pipeline_name}_{i}_{j*n_image_per_prompt+t}.webp"
+                for t in range(n_image_per_prompt)
             ]
-            for file_name, topic, prompt in zip(file_names, topics, prompts):
+            for file_name in file_names:
                 request_result.append(
                     {
                         "file_name": file_name,
-                        "topic": topic,
+                        "topic": topic_names[j],
                         "prompt": prompt,
                         "request_id": request_id,
                         "model_type": raw_pipeline_name,
@@ -63,34 +69,54 @@ def validate_pipe(
                 )
             for file_name, image in zip(file_names, images):
                 image.save(os.path.join(save_dir, file_name))
+        with open("temp.jsonl", "w") as f:
+            for item in request_result:
+                f.write(json.dumps(item) + "\n")
     return request_result
+
 
 def get_args():
     parser = ArgumentParser()
     parser.add_argument("--hf_data_repo", type=str)
     parser.add_argument("--hf_token", type=str)
+    parser.add_argument("--continue_generate", action="store_true")
 
     args = parser.parse_args()
     return args
 
+
 def main():
     args = get_args()
-    validator = Validator(prompter_cfg=config["prompter"], device=DEVICE)
+    validator = Validator(
+        prompter_cfg=config["prompter"], device=synthentic_config.DEVICE
+    )
     total_endpoints = 0
     for sd_type, endpoints in synthentic_config.ENDPOINTS.items():
         total_endpoints += len(endpoints)
     progress_bar = tqdm.tqdm(total=total_endpoints)
     metadata = []
-    with open(os.path.join(synthentic_config.SAVE_DIR, "metadata.jsonl")) as f:
-        for line in f:
-            metadata.append(json.loads(line))
-    shutil.rmtree(synthentic_config.SAVE_DIR, ignore_errors=True)
-    os.makedirs(synthentic_config.SAVE_DIR, exist_ok=True)
+    infered_endpoints = []
+    if args.continue_generate:
+        with open(os.path.join(synthentic_config.SAVE_DIR, "metadata.jsonl")) as f:
+            for line in f:
+                metadata.append(json.loads(line))
+        for item in metadata:
+            infered_endpoints.append(item["model_type"])
+        infered_endpoints = list(set(infered_endpoints))
+    else:
+        shutil.rmtree(synthentic_config.SAVE_DIR, ignore_errors=True)
+        os.makedirs(synthentic_config.SAVE_DIR, exist_ok=True)
+    print("Infered endpoints: ", infered_endpoints)
     for sd_type, endpoints in synthentic_config.ENDPOINTS.items():
         for endpoint in endpoints:
+            if endpoint in infered_endpoints:
+                continue
             if sd_type == "sd":
                 pipe = StableDiffusionPipeline.from_pretrained(
-                    endpoint, torch_dtype=torch.float16, safety_checker=None, requires_safety_checker=False
+                    endpoint,
+                    torch_dtype=torch.float16,
+                    safety_checker=None,
+                    requires_safety_checker=False,
                 )
             else:
                 pipe = StableDiffusionXLPipeline.from_pretrained(
@@ -99,7 +125,7 @@ def main():
                     use_safetensors=True,
                     variant="fp16",
                     safety_checker=None,
-                    requires_safety_checker=False
+                    requires_safety_checker=False,
                 )
             pipe.unet = torch.compile(pipe.unet, mode="reduce-overhead", fullgraph=True)
             pipe = pipe.to(synthentic_config.DEVICE)
@@ -114,7 +140,9 @@ def main():
                 negative_prompt=synthentic_config.NEGATIVE_PROMPT,
             )
             metadata.extend(sub_metadata)
-            with open(os.path.join(synthentic_config.SAVE_DIR, "metadata.jsonl"), "w") as f:
+            with open(
+                os.path.join(synthentic_config.SAVE_DIR, "metadata.jsonl"), "w"
+            ) as f:
                 for item in metadata:
                     f.write(json.dumps(item) + "\n")
             progress_bar.update(1)
@@ -129,9 +157,10 @@ def main():
     )
 
     ds.push_to_hub(
-        args.hf_data_repo, 
+        args.hf_data_repo,
         token=args.hf_token,
     )
 
-if __name__='__main__':
+
+if __name__ == "__main__":
     main()
